requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.3
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-ranged-response
    #   django-simple-captcha
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   social-auth-app-django
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-ranged-response==0.2.0
    # via django-simple-captcha
django-simple-captcha==0.5.18
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.0
    # via djoser
djoser==2.2.0
    # via -r requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pillow==10.0.0
    # via
    #   -r requirements.in
    #   django-simple-captcha
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3
    # via djangorestframework
requests==2.31.0
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
social-auth-app-django==5.2.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.4.2
    # via social-auth-app-django
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tzdata==2023.3
    # via django
urllib3==2.0.4
    # via requests
